Add files via upload
Scrupts for Accelerator Toolbox, which allows the analytical calculation of the three lowest-order in the nonlinear dispersion and the corresponding orders in the momentum compaction factor. Such scripts can be modulated to display the nonlinear dispersion, as well as the precision of the calculations, by setting the number of required slices per element of the ring.