Add files via upload
Accelerator Toolbox script, which aims at minimising the first-order momentum compaction factor of a lattice, by varying its sextupole strengths. All sextupoles are treated individually. 
INPUT : RING
OUTPUT VARIABLES :  the optimised ring, the values of the sextupoles of each generated ring, and the vector 'alpha1all' gathering the values of alpha1 of all generated rings.
OUTPUT FIGURES :  a first figure is updated along the optimisation and superimposes the second-order dispersion of all generated rings. At the end of the optimisation, three additional figures are created: the variation of alpha1 during the optimisation, the relative sextupole variation of the ring of minimum alpha1 and the comparison of the first-order dispersion of the ring of minimum alpha1 with that of the nominal ring.